# api/openapi.yaml
openapi: 3.1.1
info:
  title: CBOM-Repository Service API
  description: |-
    REST API for uploading, retrieving and searching of cryptographic software bill of materials documents.
    A cryptographic software bill of materials document adheres to CycloneDX schema version 1.6 or 1.7.

  contact:
    name: OmniTrustILM
    url: https://www.omnitrust.com
    email: getinfo@omnitrust.com

  license:
    name: MIT License
    url: https://github.com/OmniTrustILM/cbom-repository/blob/main/LICENSE.md
  version: 0.1.0

externalDocs:
  description: OmniTrustILM Documentation
  url: https://docs.omnitrust.com

paths:
  /v1/bom:
    post:
      summary: Upload a BOM
      operationId: uploadBOM
      requestBody:
        required: true
        description: >-
          CycloneDX BOM document. Select the schema version via the media-type
          `version` parameter, e.g. `application/vnd.cyclonedx+json; version=1.7`
          (supported: 1.6, 1.7). When omitted, the server's configured default
          (APP_DEFAULT_BOM_VERSION, default 1.6) is used. The document's own
          `specVersion` must match the declared version.
        content:
           application/vnd.cyclonedx+json:
            schema:
              $schema: https://json-schema.org/draft/2020-12/schema
              type: string
              format: binary

      responses:
        '201':
          description: BOM created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BOMCreateResponse'
        '400':
          description: Bad Request (e.g. malformed BOM, or unsupported/mismatched CycloneDX version)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: Conflict (e.g. BOM with same serialNumber/version already exists)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '415':
          description: >-
            Unsupported media type: the `Content-Type` is missing/blank, cannot be
            parsed, or its base type is not `application/vnd.cyclonedx+json`.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        default:
          description: General Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
    get:
      summary: Search BOMs created after a timestamp
      operationId: getBOMs
      parameters:
        - name: after
          in: query
          description: Unix timestamp
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Array of BOM entries (serialNumber + version)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BOMEntry'
        default:
          description: General Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"

  /v1/bom/{urn}:
    get:
      summary: Retrieve a BOM by URN
      operationId: getBOMByUrn
      parameters:
        - name: urn
          in: path
          description: Unique resource identifier (URN), corresponds to CycloneDX serial number.
          required: true
          schema:
            type: string
            example: "urn:uuid:3e671687-395b-41f5-a30f-a58921a69b79"
        - name: version
          in: query
          description: Optional `version`. If omitted, returns latest version.
          required: false
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Requested BOM
          content:
            application/vnd.cyclonedx+json:
              schema:
                type: string
                format: binary
        '404':
          description: BOM not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        default:
          description: General Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"

  /v1/bom/{urn}/versions:
    get:
      summary: List available BOM versions
      description: Returns all available versions of a BOM identified by its URN.
      operationId: listBomVersions
      tags:
        - BOM
      parameters:
        - name: urn
          in: path
          required: true
          description: URN of the BOM
          schema:
            type: string
      responses:
        '200':
          description: Successful response with list of BOM versions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BOMVersion'
        '400':
          description: Invalid URN supplied
        '404':
          description: BOM not found
        '500':
          description: Internal server error

  /v1/health:
    get:
      summary: Get overall health status
      operationId: getHealth
      description: |-
        Returns the health status of the service and its components.
        Status codes: 200 (UP/DEGRADED), 503 (DOWN/OUT_OF_SERVICE)
      responses:
        '200':
          description: Service is healthy or degraded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
        '503':
          description: Service is down or out of service
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /v1/health/liveness:
    get:
      summary: Kubernetes liveness probe
      operationId: getLiveness
      description: |-
        Returns liveness status. Used by Kubernetes to determine if the pod should be restarted.
        Always returns UP unless the process itself is in a failed state.
      responses:
        '200':
          description: Service is alive
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
        '503':
          description: Service is not alive
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /v1/health/readiness:
    get:
      summary: Kubernetes readiness probe
      operationId: getReadiness
      description: |-
        Returns readiness status. Used by Kubernetes to determine if the pod can accept traffic.
        Returns OUT_OF_SERVICE if critical components (e.g., storage) are unavailable.
      responses:
        '200':
          description: Service is ready to accept traffic
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
        '503':
          description: Service is not ready to accept traffic
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

components:
  schemas:

    HealthResponse:
      type: object
      description: Health status response
      properties:
        status:
          type: string
          enum: [UP, DOWN, OUT_OF_SERVICE, UNKNOWN, DEGRADED]
          description: Overall health status
          example: "UP"
        components:
          type: object
          description: Status of individual components
          additionalProperties:
            $ref: '#/components/schemas/HealthComponent'
          example:
            liveness:
              status: "UP"
            readiness:
              status: "UP"
            storage:
              status: "UP"
              details:
                latencyMs: 45
      required: [status, components]

    HealthComponent:
      type: object
      description: Health status of a single component
      properties:
        status:
          type: string
          enum: [UP, DOWN, OUT_OF_SERVICE, UNKNOWN, DEGRADED]
          description: Component health status
          example: "UP"
        details:
          type: object
          description: Additional details about the component's health
          additionalProperties: true
          example:
            latencyMs: 45
            endpoint: "https://storage.example.com"
      required: [status]

    BOMEntry:
      type: object
      required:
        - serialNumber
        - version
        - timestamp
        - cryptoStats
      properties:
        serialNumber:
          type: string
          description: BOM serial number
          example: "urn:uuid:3e671687-395b-41f5-a30f-a58921a69b79"
        version:
          type: string
          description: BOM Version - number or `original` string
          example: "1"
        timestamp:
          type: string
          format: date-time
          description: RFC 3339 timestamp when this BOM entry was created
          example: "2026-01-25T21:35:05Z"
        cryptoStats:
          $ref: '#/components/schemas/CryptoStats'

    BOMCreateResponse:
      $schema: https://json-schema.org/draft/2020-12/schema
      type: object
      description: Response returned after uploading a BOM.
      properties:
        serialNumber:
          type: string
          description: CycloneDX serial number (URN, RFC-4122).
          example: "urn:uuid:3e671687-395b-41f5-a30f-a58921a69b79"
        version:
          type: integer
          description: CycloneDX integer version number for the BOM document
          example: 1
        cryptoStats:
          $ref: '#/components/schemas/CryptoStats'
      required: [serialNumber, version, cryptoStats]
      additionalProperties: false

    BOMVersion:
      type: object
      required:
        - version
        - timestamp
        - cryptoStats
      properties:
        version:
          type: string
          example: "1"
        timestamp:
          type: string
          format: date-time
          example: "2026-01-19T21:35:05Z"
        cryptoStats:
          $ref: '#/components/schemas/CryptoStats'

    # RFC 9457 Problem Details (JSON only)
    ProblemDetails:
      $schema: https://json-schema.org/draft/2020-12/schema
      type: object
      description: Problem Details per RFC 9457 (application/problem+json).
      properties:
        type:
          type: string
          format: uri-reference
          description: A URI reference that identifies the problem type. Defaults to "about:blank".
          example: "https://example.com/problems/invalid-payload"
        title:
          type: string
          description: Short, human-readable summary of the problem type.
          example: "Invalid payload"
        status:
          type: integer
          description: HTTP status code for this occurrence of the problem.
          example: 400
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence.
          example: "The 'serialNumber' field is missing."
        instance:
          type: string
          format: uri-reference
          description: A URI reference that identifies the specific occurrence of the problem.
          example: "urn:uuid:4b96f3f7-0c2a-43f7-9c0a-7b0b6a3e2a61"
      additionalProperties: true

    CryptoStats:
      type: object
      required:
        - cryptoAssets
      properties:
        cryptoAssets:
          $ref: '#/components/schemas/CryptoAssets'

    CryptoAssets:
      type: object
      required:
        - total
        - algorithms
        - certificates
        - protocols
        - relatedCryptoMaterials
      properties:
        total:
          type: integer
          example: 1
        algorithms:
          $ref: '#/components/schemas/CryptoAssetCount'
        certificates:
          $ref: '#/components/schemas/CryptoAssetCount'
        protocols:
          $ref: '#/components/schemas/CryptoAssetCount'
        relatedCryptoMaterials:
          $ref: '#/components/schemas/CryptoAssetCount'

    CryptoAssetCount:
      type: object
      required:
        - total
      properties:
        total:
          type: integer
          example: 0